Fix SGB Borders not working on PAL SNES / SGB
fixes #45 by always adding a delay of 4 frames at the beginning of main. It seems to be enough to make loading of SGB Borders work on my PAL SNES / SGB. Adding the delay later on makes it still fail. I did not really check why. I also doubt this would really interfere with NTSC systems as there the problem did not exist and a delay of 4 frames for every game just at startup is not much. The only thing i don't know is if wait_vbl_done being called there could have some impact with other code of your engine